Step 1: Add Cocoa backend skeleton for macOS
- Define NANA_COCOA macro on macOS (replaces NANA_X11 for Cocoa backend)
- Rename old NANA_MACOS to NANA_XQUARTZ for legacy XQuartz/X11 support
- Add cocoa/platform_spec.hpp with forward declarations of X11 types
- Add platform_spec_cocoa.mm with Cocoa NSApplication initialization
- Add bedrock_cocoa.mm with Cocoa event loop stub
- Add native_window_interface_cocoa.cpp with window operation stubs
- Add cocoa/msg_dispatcher.hpp for Cocoa message dispatching
- Update CMakeLists.txt: add OBJCXX language, cocoa source dirs
- Update OS.cmake: link Cocoa/CoreGraphics/CoreText frameworks
- Fix msg_packet.hpp: conditional X11 includes for Cocoa compatibility
- Fix platform_spec_selector.hpp: select cocoa header on NANA_COCOA
- Fix platform_abstraction.cpp: Cocoa font stubs
- Fix thread_t conflict with macOS mach/mach_types.h
- Guard X11-only code paths with #ifndef NANA_COCOA
- Exclude X11-specific source files from Cocoa build

+++ source/system/platform.cpp
@@ -102,10 +102,10 @@ namespace system
 		return (thread_t)::GetCurrentThreadId();
 #elif defined(NANA_LINUX)
 		return (thread_t)::syscall(__NR_gettid);
-#elif defined(NANA_MACOS)
+#elif defined(NANA_XQUARTZ)
 		return (thread_t)::syscall(SYS_thread_selfid);
 #elif defined(NANA_POSIX)
-        return (thread_t)pthread_self();
+        return (thread_t)(uintptr_t)pthread_self();
 #endif
 	}
